refactor empty_string for python3
Shifted test to unit dir

Added tabs
