+++ app/api/helpers/utilities.py
@@ -26,12 +26,7 @@ def require_relationship(resource_list, data):
 
 
 def string_empty(value):
-    is_not_str_type = type(value) is not str
-    if sys.version_info[0] < 3:
-        is_not_str_type = is_not_str_type and type(value) is not unicode
-    if type(value) is not value and is_not_str_type:
-        return False
-    return not (value and value.strip() and value != u'' and value != u' ')
+    return isinstance(value, str) and not value.strip()
 
 
 def strip_tags(html):
